Adding missing test of my last commit and moving custom csv UnicodeWriter to compat.py module
